Perform Suggested Changes By @Rushikesh-josh regarding Path, Functional Components, Containers
