+++ react-assignment-1/src/types/todo.ts
@@ -1,0 +1,5 @@
+export type Todo = {
+    id: number,
+    title: string,
+    done: "Complete" | "Pending"
+}
